// eshop-backoffice/src/main/resources/backoffice/src/app/core/enums/enums.ts
export enum AffiliateEnum {
    TECHNO_CH = 'technomag',
    DEREND_CH = 'derendinger-ch',
    DEREND_AT = 'derendinger-at',
    MATIK_AT = 'matik-at'
}

export enum AFFILIATE_CODE {
    AT =  'at',
    CH = 'ch',
    AXCZ = 'axcz',
    CZ = 'cz'
};

export enum BUILD_ENVIRONMENT {
    PRODUCTION = '',
    PRE_PRODUCTION = 'Pre-Production',
    TEST = 'Test',
    DEVELOPMENT = 'Development',
    FEATURE = 'Feature'
}

export enum messageLocationTypeEnum {
    AFFILIATE = 'AFFILIATE',
    CUSTOMER = 'CUSTOMER'
}

export enum messageTypeEnum {
    BANNER = 'BANNER',
    PANEL = 'PANEL'
}

export enum SORT_TYPE {
    ASC, DESC
}

export enum GENDER {
    MALE = 'Male',
    FEMALE = 'Female'
}

export enum CURRENCY {
    SWISS_CURRENCY = 'SFr.',
    EURO_CURRENCY = 'EUR',
    CHF_CURRENCY = 'CHF'
}

export enum KEY_BOARD {
    ENTER = 13,
    ARROW_DOWN = 40,
    ARROW_UP = 38,
    ZERO = 48,
    NUMBER_NINE = 57,
    SEPARATOR_COMMAS = 44,
    SEPARATOR_DOT = 46,
    BACK_SPACE = 8,
    ARROW_LEFT = 37,
    ARROW_RIGHT = 39,
    HOME = 36,
    END = 35,
}

export enum CHARACTER {
    COMMAS = ',',
    DOT = '.',
    APOSTROPHE = '\''
}

export enum CUSTOMER_SEARCH_MODE {
    CUSTOMER = 'customer',
    LICENCES = 'licences',
}